GRP-2121: SQL provisioning (commit 2)

+++ grouper-misc/grouperClient/src/java/edu/internet2/middleware/grouperClient/jdbc/tableSync/GcTableSync.java
@@ -254,7 +254,7 @@ public class GcTableSync {
         String sql = "select " + this.tableMetadata.getRealTimeLastUpdatedColumnMetadata().getColumnName() + " from " + this.tableMetadata.getSchemaFrom() + "." + this.getRealTimeTable() + " where " 
             + realTimeLastUpdatedColumnMetadata.getColumnName() + " > ?";
   
-        int total = new GcDbAccess().connectionName(this.databaseFrom).sql(sql).callbackResultSet(new GcResultSetCallback() {
+        int total = (Integer)new GcDbAccess().connectionName(this.databaseFrom).sql(sql).callbackResultSet(new GcResultSetCallback() {
   
           @Override
           public Object callback(ResultSet resultSet) throws Exception {
